% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp_ar1_mae.R
\name{dgp_ar1_mae}
\alias{dgp_ar1_mae}
\title{Generate data according to Equation 1 page 183 of Enders.}
\usage{
dgp_ar1_mae(periods = 100, sd = 1, gamma = 1, theta = 0.5)
}
\arguments{
\item{periods}{A time series object}

\item{sd}{Standard deviation}

\item{gamma}{Gamma}

\item{theta}{Theta}
}
\value{
A data frame
}
\description{
Returns a vector of time series values
}
\examples{
dgp_ar1_mae(periods = 100, sd = 1, gamma = 1, theta = .5)
}
\author{
Gary Cornwall and Jeffrey Chen
}
