% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp_arma.R
\name{dgp_arma}
\alias{dgp_arma}
\title{Generate data according to ARMA process}
\usage{
dgp_arma(periods = 100, sd = 1, gamma = 1, theta = -0.25)
}
\arguments{
\item{periods}{A time series object}

\item{sd}{Standard deviation}

\item{gamma}{Unit root}

\item{theta}{Theta}
}
\value{
A data frame
}
\description{
Returns a time series vector
}
\examples{
dgp_arma(periods = 100,
         sd = 1,
         gamma = 1,
         theta = -.25)
}
\author{
Gary Cornwall and Jeffrey Chen
}
