% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp_enders1.R
\name{dgp_enders1}
\alias{dgp_enders1}
\title{Generate time series with properties according to Equation 1 page 183 of Enders (2008).}
\usage{
dgp_enders1(periods = 100, sd = 1, alpha0 = 1, alpha2 = 0.005, gamma = 1)
}
\arguments{
\item{periods}{Number of periods}

\item{sd}{Standard deviation}

\item{alpha0}{Alpha0}

\item{alpha2}{Alpha2}

\item{gamma}{Gamma}
}
\value{
A data frame
}
\description{
Returns a vector of time series values
}
\examples{
dgp_enders1(periods = 200, sd = 1, alpha = 1, alpha2 = 0.005, gamma = 1)
}
\references{
Enders, Walter. 2008 Applied econometric time series. New York: Wiley.
}
\author{
Gary Cornwall and Jeffrey Chen
}
