% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp_enders3.R
\name{dgp_enders3}
\alias{dgp_enders3}
\title{Generate time series with properties according to Equation 3 page 183 of Enders (2008).}
\usage{
dgp_enders3(periods = 100, sd = 1, gamma = 1)
}
\arguments{
\item{periods}{Number of periods}

\item{sd}{Standard deviation}

\item{gamma}{Gamma}
}
\value{
A data frame
}
\description{
Returns a vector of time series values
}
\examples{

dgp_enders3(periods = 200,
            sd = 1,
           gamma = 1)
}
\references{
Enders, Walter. 2008 Applied econometric time series. New York: Wiley.
}
\author{
Gary Cornwall and Jeffrey Chen
}
