% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp_engle.R
\name{dgp_engle}
\alias{dgp_engle}
\title{Generate data according to Heteroskedasticitic error term from Engle (1982)}
\usage{
dgp_engle(
  periods = 100,
  sd = 1,
  gamma = 1,
  theta = 0.5,
  alpha0 = 1,
  alpha1 = 0.5
)
}
\arguments{
\item{periods}{A time series object}

\item{sd}{Standard deviation}

\item{gamma}{Gamma}

\item{theta}{Theta}

\item{alpha0}{Alpha0}

\item{alpha1}{Alpha1}
}
\value{
A data frame
}
\description{
Returns a vector of
}
\examples{
dgp_engle(periods = 100, sd = 1, gamma = 1, theta = .5, alpha0 = 1, alpha1 = .5)
}
\references{

}
\author{
Gary Cornwall and Jeffrey Chen
}
