% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_features.R
\name{gen_features}
\alias{gen_features}
\title{Estimate a battery of statistical tests for unit roots and time series features}
\usage{
gen_features(bank, cval = 0.05, run_par = TRUE)
}
\arguments{
\item{bank}{A list of time series objects constructed by gen_bank function}

\item{cval}{A critical value as a p-value}

\item{run_par}{Boolean indicating whether to compute in parallel.}
}
\value{
Returns a list object of time series objects with their attributes.

A data frame
}
\description{
Returns a data frame of unit root tests and time series features
}
\details{
Returns a set of commonly used time series statistical tests and time series features.
}
\examples{
#Create simulated time series using Enders 1
new_series <- gen_bank(iter = 300,
                       dgp_params = list(dgp = "dgp_enders1",
                                         periods = 100,
                                         sd = 1,
                                         alpha0 = 1,
                                         alpha2 =  .005,
                                         gamma = 1))
#Generate features
feat <- gen_features(bank = new_series)
}
\author{
Gary Cornwall and Jeffrey Chen
}
