% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_test.R
\name{gen_test}
\alias{gen_test}
\title{Construct ML models for use as a ML test}
\usage{
gen_test(
  train_set,
  val_set,
  model_params = list(list(method = "ranger", cv_folds = 5, tune_grid =
    expand.grid(mtry = c(3, 9, 27), splitrule = "gini", min.node.size = c(2, 4, 8))),
    list(method = "xgbTree", cv_folds = 5, tune_grid = expand.grid(nrounds = 300, eta =
    c(0.01, 0.03, 0.1, 0.3, 0.5), gamma = 0, colsample_bytree = c(0.8, 1),
    min_child_weight = 1, subsample = c(0.8, 1), max_depth = c(4, 6))))
)
}
\arguments{
\item{train_set}{A training set outputted from gen_features function}

\item{val_set}{A validation set outputted from gen_features function to be used to calibrate threshold}

\item{model_params}{A list of parameters for specifying models. Multiple methods allowed. Requires a 'method' tag to specify algorithm, 'cv_folds' to indicate number of folds for cross validation, then data frame of hyperparameters for inclusion in tune_grid.}
}
\value{
A list object
}
\description{
Returns a list of model objects with accuracies and ROC coordinates based on test set
}
\examples{
gen_calc(train_set, val_set, model_params)
}
\author{
Gary Cornwall and Jeffrey Chen
}
