% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_noise.R
\name{make_noise}
\alias{make_noise}
\title{Draw random values from an assortment of distributions}
\usage{
make_noise(
  n = 1,
  mu = 0,
  sd = 1,
  hi_lo = 1,
  error_type = "gaussian",
  symmetric = TRUE
)
}
\arguments{
\item{n}{Number of observations Default = 1.}

\item{mu}{Mean. Default = 0.}

\item{sd}{Standard deviation. Default = 1.}

\item{hi_lo}{A value to represent maximum value and minimum value for binomial distribution. or uniform distribution. A value of 2 will set a maximum of +2 and a minimum of -2. Default = 1.}

\item{error_type}{Error distribution. Options include "gaussian", "binomial" and "uniform".}

\item{symmetric}{Boolean indicating whether simulate symmetry. Default = TRUE. When FALSE, an absolute value is applied to the noise parameter to force assymmetry}
}
\value{
A numeric value or vector
}
\description{
Draw random values from an assortment of distributions
}
\examples{
#Draw 100 from a binomial distribution of -3 and 3
make_noise(n = 100, error_type = "binomial", hi_lo = 3)

#Draw 1 value from a Gaussian distribution N(2,10)
make_noise(n = 1, error_type = "gaussian", mu = 2, sd = 10)
}
\author{
Gary Cornwall and Jeffrey Chen
}
