% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_test.R
\name{ml_test}
\alias{ml_test}
\title{Apply ML-based hypothesis test to a bank of time series}
\usage{
ml_test(
  bank,
  pvalue = 0.05,
  custom_model,
  original = TRUE,
  verdicts = 1,
  run_par = TRUE,
  num_cores = -9999,
  fanfare = FALSE
)
}
\arguments{
\item{bank}{List object with at least one time series}

\item{pvalue}{A numeric value indicating the desired p-value. (Default = 0.05) To retrieve the best accuracy, enter "-9999".}

\item{custom_model}{A unitrootML model object with a custom-trained model. Note that a custom model is used only when original is set to TRUE and a unitrootML model object is supplied. (Default = NULL)}

\item{original}{Boolean indicating whether to use default model or a custom model. The default model is a gradient boost model with a sensitivity rated at 0.924 and specificity rated at 0.952. (Default = TRUE)}

\item{verdicts}{Integer indicating whether to (1) report only ML model verdicts, (2) both ML and test statistic verdicts based on thresholds calibrated from selected cost ratio. (Default = 1)}

\item{run_par}{Boolean indicating whether to compute in parallel (Default = TRUE).}

\item{num_cores}{Number of logical cores to use for parallel processing. -9999 indicates maximum allowed minus one. Otherwise, provide an integer.  (Default = -9999).}

\item{fanfare}{Boolean indicating whether to print results (Default = FALSE).}
}
\value{
A unitrootML object containing test results
}
\description{
Apply ML-based hypothesis test to a bank of time series
}
\examples{
ml_test(list(ts(rnorm(120, 10,10), freq=12)), pvalue = 0.05)
}
\author{
Gary Cornwall and Jeffrey Chen
}
