% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_test_all.R
\name{ml_test_all}
\alias{ml_test_all}
\title{Apply ML-based hypothesis test to a bank of time series}
\usage{
ml_test_all(
  bank,
  costratio = 1,
  original = TRUE,
  custom_model,
  run_par = TRUE,
  fanfare = FALSE
)
}
\arguments{
\item{bank}{List object with at least one time series}

\item{costratio}{A numeric cost ratio (default = 1)}

\item{original}{Boolean indicating whether to use pre-trained models from Cornwall and Chen (2020).}

\item{custom_model}{hypML model object with a custom-trained model}

\item{run_par}{Boolean indicating whether to compute in parallel.}

\item{fanfare}{Boolean indicating whether to print results.}
}
\value{
A HypML object containing test results
}
\description{
Apply ML-based hypothesis test to a bank of time series
}
\examples{
ml_test(ts(rnorm(120, 10,10), freq=12))
}
\author{
Gary Cornwall and Jeffrey Chen
}
