% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_01.R
\name{scale_01}
\alias{scale_01}
\title{Scale data from (0 to 1) to (0 to 1)}
\usage{
scale_01(x, a, b)
}
\arguments{
\item{x}{A time series object}

\item{a}{param a}

\item{b}{param b}
}
\value{
Scaled values
}
\description{
Scaling to 0/1 from 0/1. This is a support function that is not intended for the end user to directly use.
}
\examples{
scale_01(ts(rnorm(120, 10,10),
           frequency=12))
}
\references{
https://robjhyndman.com/hyndsight/tscharacteristics/
}
\author{
Rob Hyndman
}
