% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_inf.R
\name{scale_inf}
\alias{scale_inf}
\title{Scale a series from (0 to infinity) to (0 to 1)}
\usage{
scale_inf(x, a = 1.51, b = 5.993)
}
\arguments{
\item{x}{A time series object}

\item{a}{param a}

\item{b}{param b}
}
\value{
Scaled values
}
\description{
Scaling from (0 to infinity) to (0 to 1). This is a support function that is not intended for the end user to directly use.
}
\examples{
scale_inf(ts(rnorm(120, 10,10), frequency=12))
}
\references{
https://robjhyndman.com/hyndsight/tscharacteristics/
}
\author{
Rob Hyndman
}
