% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_decomp.R
\name{seas_decomp}
\alias{seas_decomp}
\title{Extract seasonal decomposition from a time series}
\usage{
seas_decomp(x, transform = TRUE)
}
\arguments{
\item{x}{A time series object}

\item{transform}{Boolean for whether a Box-Cox transform is applied}
}
\value{
A list object with QS-statistic and critical value
}
\description{
Test for seasonality in a time series based on lags
}
\examples{
seas_decomp(ts(rnorm(120, 10,10), frequency=12))
}
\references{
https://robjhyndman.com/hyndsight/tscharacteristics/
}
\author{
Rob Hyndman
}
