% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_summary.R
\name{testing_summary}
\alias{testing_summary}
\title{Return summary of ML-based hypothesis test}
\usage{
testing_summary(hypML_obj)
}
\arguments{
\item{hypML_obj}{A hypML object from ml_test function}
}
\value{
Retrieve test results
}
\description{
Print summary of ML-based hypothesis test
}
\examples{

#Create a list of time series
x <- list(norm1 = ts(rnorm(120, 10,10), freq=12),
          with_trend = ts(rnorm(120, 10,10) + 1:120, freq=12),
          uroot = ts(cumsum(rnorm(120,10,10)), freq = 12))

#Screen the time series for unit roots using the default model
output <- ml_test(bank = x,
                 pvalue = 0.05)

#Retrieve test summary
testing_summary(hypML_obj = output)


}
\author{
Gary Cornwall and Jeffrey Chen
}
