% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_calc.R
\name{threshold_calc}
\alias{threshold_calc}
\title{Calculate threshold given cost ratios}
\usage{
threshold_calc(roc_obj, costratio)
}
\arguments{
\item{roc_obj}{A training set outputted from gen_features function}

\item{costratios}{A validation set outputted from gen_features function to be used to calibrate threshold}
}
\value{
A list object
}
\description{
Returns a list of model objects with accuracies and ROC coordinates based on test set
}
\examples{
threshold_calc(roc_obj, costratio)
}
\author{
Gary Cornwall and Jeffrey Chen
}
