% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_features.R
\name{ts_features}
\alias{ts_features}
\title{Estimate a battery of statistical tests for unit roots and time series features}
\usage{
ts_features(series, cval = 0.05)
}
\arguments{
\item{series}{A time series object}

\item{cval}{A critical value as a p-value}
}
\value{
A data frame
}
\description{
Returns a set of commonly used time series statistical tests and time series features.
}
\examples{
ts_features(ts(rnorm(120, 10,10), freq=12))
}
\author{
Gary Cornwall and Jeffrey Chen
}
