% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_features.R
\name{gen_features}
\alias{gen_features}
\title{Estimate a battery of statistical tests for unit roots and time series features}
\usage{
gen_features(bank, cval = 0.05, run_par = TRUE)
}
\arguments{
\item{bank}{A list of time series objects constructed by gen_bank function}

\item{cval}{A critical value as a p-value}

\item{run_par}{Boolean indicating whether to compute in parallel.}
}
\value{
Returns a list object of time series objects with their attributes.

A data frame
}
\description{
Returns a data frame of unit root tests and time series features
}
\details{
Returns a set of commonly used time series statistical tests and time series features.
}
\examples{
gen_features(bank)
}
\author{
Gary Cornwall and Jeffrey Chen
}
